Introduce (nonstandard) annotation for fields that represent sequences
The type annotation 'seq' is a hint to systems using the parsed type
that, despite the size of the type, the type's values come from
a sequential generator.

A `uint256 seq foo` acts in all other ways like a `uint256 foo`. But
the values of the `uint256 seq foo` are "expected" to be small. Thus,
systems storing or indexing decoded `foo` values may *default* to a
smaller storage representation for foo (e.g. a 64-bit field) until
a decoded value larger than that representation is encountered.

`seq` is not a guarantee that the values will always be small. Systems
that rely on `seq` for space-efficiency, must either be capable of
migrating their data to the literal storage representation specified
by the type, at any time; or must enter an terminal invalid state when
such data is encountered.

Silently casting larger values to the smaller storage representation is
incorrect, and will result in undefined behavior. Implementations will
not be considered "`seq`-conformant" if they do this.